requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
Brotli==1.1.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
cssselect2==0.8.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.14
fonttools==4.58.4
h11==0.16.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
motor==3.7.1
passlib==1.7.4
pillow==11.3.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
pymongo==4.13.2
pyphen==0.17.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
weasyprint==65.1
webencodings==0.5.1
websockets==15.0.1
zopfli==0.2.3.post1
